downloader: rename twitch.py to providers.py
Since it will now contain non-twitch providers also.
